refactor: http app and handlers moved to package controllers
http App and all methods for App to app_handlers

all tests passed
